fix(content-explorer): Request fields for Preview SDK file cache
