+++ src/utils/__tests__/fields.test.js
@@ -1,5 +1,4 @@
 import {
-    CONTENT_EXPLORER_FOLDER_FIELDS_TO_FETCH,
     FOLDER_FIELDS_TO_FETCH,
     PREVIEW_FIELDS_TO_FETCH,
     SIDEBAR_FIELDS_TO_FETCH,
@@ -81,33 +80,6 @@ describe('util/fields', () => {
             FIELD_IS_DOWNLOAD_AVAILABLE,
             FIELD_REPRESENTATIONS,
             FIELD_URL,
-        ]);
-    });
-
-    test('should fetch correct folder fields for the Content Explorer', () => {
-        expect(CONTENT_EXPLORER_FOLDER_FIELDS_TO_FETCH).toEqual([
-            FIELD_ID,
-            FIELD_NAME,
-            FIELD_TYPE,
-            FIELD_SIZE,
-            FIELD_PARENT,
-            FIELD_EXTENSION,
-            FIELD_PERMISSIONS,
-            FIELD_PATH_COLLECTION,
-            FIELD_MODIFIED_AT,
-            FIELD_CREATED_AT,
-            FIELD_MODIFIED_BY,
-            FIELD_HAS_COLLABORATIONS,
-            FIELD_IS_EXTERNALLY_OWNED,
-            FIELD_ITEM_COLLECTION,
-            FIELD_AUTHENTICATED_DOWNLOAD_URL,
-            FIELD_IS_DOWNLOAD_AVAILABLE,
-            FIELD_REPRESENTATIONS,
-            FIELD_URL,
-            FIELD_FILE_VERSION,
-            FIELD_SHA1,
-            FIELD_SHARED_LINK,
-            FIELD_WATERMARK_INFO,
         ]);
     });
